Recommend using shoulda-matchers
It's cumbersome to manually test validations and ActiveModel relations,
and there's no good reason not to use a mature third-party gem that
provides convenient matchers that do that.

It's quite fair to recommend shoulda-matchers, and it's the most known
third-party RSpec matcher gem since we do already recommend FactoryBot
and TimeCop.

Relates to #33