Reorganize `let` recommendation
The two recommendations were overlapping in a point that `let` should be
preferred to local variables, and the fact that `let`s should be used
when their values are to be shared across several examples was somewhere
between the lines. Also, the recommendation when to use `let!` was in
the very end. It wasn't mentioned that it doesn't make much sense, and
is even bad for readability to overuse `let`s for primitive values, e.g.

    let(:cost) { 30 }

where it can be:

    it '...' do
      cost = 30
      ...
    end

    it '...' do
      cost = 30
      ...
    end

Kept only one anchor, of the recommendation that was introduced earlier.