Add Declare Constants guideline
Constants defined in a block are defined in a global namespace, and are
not cleared up between examples.

If several examples may define a `DummyClass`, instead of being a blank
slate class as it will be in the first example, subsequent examples will
be reopening it and modifying its behaviour in unpredictable ways.  Even
worse when a class that exists in the codebase is reopened.

Anonymous classes are fine, since they don't result in global namespace
name clashes.