Add Implicit Block Expectation guideline
Implicit syntax is discouraged to use by RSpec Core team and the
majority of voters.

There were no good arguments for using the syntax except for brevity and
avoiding repetition, but there are better options to achieve the same
goal, e.g. by extracting the lengthy block to methods (instead of
putting it inside the lambda).

Fixes #76

https://www.reddit.com/r/ruby/comments/cejl3q/call_for_discussion_rspec_implicit_block/
https://blog.rubystyle.guide/rspec/2019/07/17/rspec-implicit-block-syntax.html
https://github.com/rubocop-hq/rspec-style-guide/issues/76